requirements.txt
Django==4.2.8
django-dbml==0.7.0
django-environ==0.11.2
django-extensions==3.2.3
django-filter==23.5
django-phonenumber-field==7.3.0
django-phonenumbers==1.0.1
django-sesame==3.2.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
gunicorn==21.2.0
idanalyzer==1.2.2
markdown==3.5.1
pre-commit==3.6.0
psycopg2-binary==2.9.9
pydot==2.0.0
pygraphviz==1.11
pyparsing==3.1.1
pytest==7.4.4
pytest-django==4.7.0
python-jose[cryptography]==3.3.0
requests==2.31.0
sendgrid==6.11.0
sendgrid==6.11.0
# payments
stripe==7.12.0
whitenoise==6.6.0
